fix: wrapped component into suspense
